+++ includes/sandbox.php
@@ -1,35 +1,13 @@
 <?php
 
-/*
-?>
-<script type="text/javascript">
-jQuery(function($){
-    $(".accordion").accordion({ header: "h3" });
-    $(".accordion").last().accordion("option", "icons", false);
-});
-</script>
-
-<div class="accordion">
-    <div>
-        <h3><a href="#">First</a></h3>
-        <div>Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet.</div>
-    </div>
-    <div>
-        <h3><a href="#">Second</a></h3>
-        <div>Phasellus mattis tincidunt nibh.</div>
-    </div>
-    <div>
-        <h3><a href="#">Third</a></h3>
-        <div>Nam dui erat, auctor a, dignissim quis.</div>
-    </div>
-</div>
-<?php
-*/
-
 // Example to get settings
-// $my_settings = get_option( 'dsbl_menu_settings' );
+// $my_settings = get_option( 'dsbl_general_settings' );
 
 // echo '<pre>'; print_r( $my_settings ); echo '</pre>';
+
+// if ( $my_settings['disable_posts'] == 'disable' ) {
+//     echo 'yesy';
+// }
 
 // if ( is_array( $my_settings ) || is_object( $my_settings ) ) {
 //     foreach ( $my_settings as $group => $item ) {